Added cleaned-up and partially processed data and source code to generate it.
Beginning of refactoring: I added a new program that reads the raw
data, rearranges into three lists of dicts (one for users, one for
posts, one for comments) and savers them to .csv files. I also added
the files.

Adds a procedure that discovers users who wrote posts but not comments
(which tranlsates into nodes with some in-degree but zero out-degree)